+++ mmif-python/mmif/serialize/model.py
@@ -1,10 +1,24 @@
+"""
+The :mod:`model` module contains the classes used to represent an
+abstract MMIF object as a live Python object.
+
+The :class:`MmifObject` class or one of its derivatives is subclassed by
+all other classes defined in this SDK, except for :class:`MmifObjectEncoder`.
+
+These objects are generally instantiated from JSON, either as a string
+or as an already-loaded Python dictionary. This base class provides the
+core functionality for deserializing MMIF JSON data into live objects
+and serializing live objects into MMIF JSON data. Specialized behavior
+for the different components of MMIF is added in the subclasses.
+"""
+
 import logging
 import json
 from pyrsistent import pvector, m, pmap, s, PVector, PMap, PSet, thaw
 from datetime import datetime
 
 from deepdiff import DeepDiff
-from typing import Union, Any, Dict, Optional, TypeVar, Generic
+from typing import Union, Any, Dict, Optional, TypeVar, Generic, Type, Generator, Iterator
 
 T = TypeVar('T')
 
@@ -22,67 +36,88 @@ __all__ = [
 class MmifObject(object):
     """
     Abstract superclass for MMIF related key-value pair objects.
+
+    Any MMIF object can be initialized as an empty placeholder or
+    an actual representation with a JSON formatted string or equivalent
+    `dict` object argument.
+
+    This superclass has two specially designed instance variables, and these
+    variable names cannot be used as attribute names for MMIF objects.
+
+    1. _unnamed_attributes
+       only can be either None or an empty dictionary. If it's set to None,
+       it means the class won't take any ``Additional Attributes`` in the JSON
+       schema sense. If it's a dict, users can throw any k-v pairs to the
+       class, EXCEPT for the reserved two key names.
+    2. _attribute_classes:
+       this is a dict from a key name to a specific python class to use for
+       deserialize the value. Note that a key name in this dict does NOT
+       have to be a *named* attribute, but is recommended to be one.
+
+    # TODO (krim @ 8/17/20): this dict is however, a duplicate with the type hints in the class definition.
+    Maybe there is a better way to utilize type hints (e.g. getting them as a programmatically), but for now
+    developers should be careful to add types to hints as well as to this dict.
+
+    Also note that those two special attributes MUST be set in the __init__()
+    before calling super method, otherwise deserialization will not work.
+
+    And also, a subclass that has one or more *named* attributes, it must
+    set those attributes in the __init__() before calling super method. When
+    serializing a MmifObject, all *empty* attributes will be ignored, so for
+    optional named attributes, you must leave leave the values empty, but
+    NOT None. Any None-valued named attributes will cause issues with current
+    implementation.
+
+    :param mmif_obj: JSON string or `dict` to initialize an object.
+     If not given, an empty object will be initialized, sometimes with
+     an ID value automatically generated, based on its parent object.
     """
+    
     reserved_names: PSet = s('reserved_names', '_unnamed_attributes', '_attribute_classes')
     _unnamed_attributes: Optional[dict]
     _attribute_classes: PMap = m()  # Mapping: str -> Type
 
     def __init__(self, mmif_obj: Union[str, dict] = None) -> None:
-        """
-        Any MMIF object can be initialized as an empty placeholder or
-        an actual representation with a JSON formatted string or equivalent
-        `dict` object argument.
-
-        This superclass has two specially designed instance variables, and these
-        variable names cannot be used as attribute names for MMIF objects.
-        1. _unnamed_attributes
-          only can be either None or an empty dictionary. If it's set to None,
-          it means the class won't take any ``Additional Attributes`` in the JSON
-           schema sense. If it's a dict, users can throw any k-v pairs to the
-           class, EXCEPT for the reserved two key names.
-        2. _attribute_classes:
-          this is a dict from a key name to a specific python class to use for
-          deserialize the value. Note that a key name in this dict does NOT
-          have to be a *named* attribute, but is recommended to be one.
-        # TODO (krim @ 8/17/20): this dict is however, a duplicate with the type hints in the class definition.
-        Maybe there is a better way to utilize type hints (e.g. getting them as a programmatically), but for now
-        developers should be careful to add types to hints as well as to this dict.
-
-        Also note that those two special attributes MUST be set in the __init__()
-        before calling super method, otherwise deserialization will not work.
-
-        And also, a subclass that has one or more *named* attributes, it must
-        set those attributes in the __init__() before calling super method. When
-        serializing a MmifObject, all *empty* attributes will be ignored, so for
-        optional named attributes, you must leave leave the values empty, but
-        NOT None. Any None-valued named attributes will cause issues with current
-        implementation.
-
-        :param mmif_obj: JSON string or `dict` to initialize an object.
-         If not given, an empty object will be initialized, sometimes with
-         an ID value automatically generated, based on its parent object.
-        """
         if not hasattr(self, '_unnamed_attributes'):
             self._unnamed_attributes = {}
         if mmif_obj is not None:
             self.deserialize(mmif_obj)
 
     def disallow_additional_properties(self) -> None:
+        """
+        Call this method in :func:`__init__` to prevent the insertion
+        of unnamed attributes after initialization.
+        """
         self._unnamed_attributes = None
 
-    def _named_attributes(self):
+    def _named_attributes(self) -> Generator[str, None, None]:
+        """
+        Returns a generator of the names of all of this object's named attributes.
+
+        :return: generator of names of all named attributes
+        """
         return (n for n in self.__dict__.keys() if n not in self.reserved_names)
 
     def serialize(self, pretty: bool = False) -> str:
         """
-        Generates JSON-LD representation of an object.
+        Generates JSON representation of an object.
 
         :param pretty: If True, returns string representation with indentation.
-        :return: JSON-LD string of the object.
+        :return: JSON string of the object.
         """
         return json.dumps(self._serialize(), indent=2 if pretty else None, cls=MmifObjectEncoder)
 
     def _serialize(self, alt_container: Dict = None) -> Union[None, dict]:
+        """
+        Maps a MMIF object to a plain python dict object,
+        rewriting internal keys that start with '_' to
+        start with '@' per the JSON-LD schema.
+
+        If a subclass needs special treatment during the mapping, it needs to
+        override this method.
+
+        :return: the prepared dictionary
+        """
         container = alt_container if alt_container is not None else self._unnamed_attributes
         serializing_obj = {}
         try:
@@ -123,7 +158,7 @@ class MmifObject(object):
     @staticmethod
     def _load_json(json_obj: Union[dict, str]) -> dict:
         """
-        Maps JSON-LD-format MMIF strings and dicts into Python dicts
+        Maps JSON-format MMIF strings and dicts into Python dicts
         with identifier-compliant keys. To do this, it replaces "@"
         signs in JSON-LD field names with "_" to be python-compliant.
 
@@ -132,8 +167,8 @@ class MmifObject(object):
         >>> "_value" in MmifObject._load_json('{ "@type": "some_type", "@value": "some_value"}').keys()
         True
 
-        :param json_str:
-        :return:
+        :param json_str: the JSON data to load and process
+        :return: the mapped data as a dict
         """
         def from_atsign(d: Dict[str, Any]) -> dict:
             for k in list(d.keys()):
@@ -173,10 +208,13 @@ class MmifObject(object):
         Maps a plain python dict object to a MMIF object.
         If a subclass needs special treatment during the mapping, it needs to
         override this method.
+
         This defalt method won't work for generic types (e.g. List[X], Dict[X, Y]).
         For now, lists are abstracted as DataList and dicts are abstracted as XXXMedata classes.
         However, if an attribute uses a generic type (e.g. view_metadata.contains: Dict[str, Contain])
         that class should override _deserialize of its own.
+
+        :param input_dict: the prepared JSON data that defines the object
         """
         for k, v in input_dict.items():
             if self._attribute_classes and k in self._attribute_classes:
@@ -330,7 +368,7 @@ class MmifObjectEncoder(json.JSONEncoder):
 
     def default(self, obj: 'MmifObject'):
         """
-        Overrides default encoding behavior to prioritize :func: MmifObject.serilize() .
+        Overrides default encoding behavior to prioritize :func:`MmifObject.serialize()`.
         """
         if hasattr(obj, '_serialize'):
             return obj._serialize()
@@ -343,6 +381,13 @@ class MmifObjectEncoder(json.JSONEncoder):
 
 
 class DataList(MmifObject, Generic[T]):
+    """
+    The DataList class is an abstraction that represents the
+    various lists found in a MMIF file, such as media, submedia,
+    views, and annotations.
+
+    :param Union[str, list] mmif_obj: the data that the list contains
+    """
     def __init__(self, mmif_obj: Union[str, list] = None):
         self.reserved_names = self.reserved_names.add('_items')
         self._items: Dict[str, T] = dict()
@@ -352,9 +397,17 @@ class DataList(MmifObject, Generic[T]):
         super().__init__(mmif_obj)
 
     def _serialize(self, *args, **kwargs) -> list:
+        """
+        Internal serialization method. Returns a list.
+
+        :return: list of the values of the internal dictionary.
+        """
         return list(super()._serialize(self._items).values())
 
     def deserialize(self, mmif_json: Union[str, list]) -> None:
+        """
+        Passes the input data into the internal deserializer.
+        """
         if isinstance(mmif_json, str):
             mmif_json = json.loads(mmif_json)
         self._deserialize(mmif_json)
@@ -363,12 +416,35 @@ class DataList(MmifObject, Generic[T]):
         raise NotImplementedError()
 
     def get(self, key: str) -> Optional[T]:
+        """
+        Standard dictionary-style get() method, albeit with no ``default``
+        parameter. Relies on the implementation of __getitem__.
+
+        Will return ``None`` if the key is not found.
+
+        :param key: the key to search for
+        :return: the value matching that key
+        """
         try:
             return self[key]
         except KeyError:
             return None
 
     def _append_with_key(self, key: str, value: T, overwrite=False) -> None:
+        """
+        Internal method for appending a key-value pair. Subclasses should
+        implement an append() method that extracts a key from the list data
+        or generates a key programmatically (such as an index), depending
+        on the data type.
+
+        :param key: the desired key to append
+        :param value: the value associated with the key
+        :param overwrite: if set to True, will overwrite an existing K-V pair
+         if the key already exists. Otherwise, raises a KeyError.
+        :raise KeyError: if ``overwrite`` is False and the ``key`` is already
+         present in the DataList.
+        :return: None
+        """
         if not overwrite and key in self._items:
             raise KeyError(f"Key {key} already exists")
         else:
@@ -389,16 +465,16 @@ class DataList(MmifObject, Generic[T]):
         else:
             super().__setitem__(key, value)
 
-    def __iter__(self):
+    def __iter__(self) -> Iterator[T]:
         return self._items.values().__iter__()
 
-    def __len__(self):
+    def __len__(self) -> int:
         return self._items.__len__()
 
-    def __reversed__(self):
+    def __reversed__(self) -> Iterator[T]:
         return reversed(list(self._items.values()))
 
-    def __contains__(self, item):
+    def __contains__(self, item) -> bool:
         return item in self._items